third_party/WebKit/LayoutTests/ambient-light/add-listener.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/js-test.js"></script>
<script>
description('Tests that adding a new devicelight event listener invokes a callback.');

var mockLight = 10;

if (window.testRunner)
    testRunner.setMockDeviceLight(mockLight);
else
    debug('This test can not be run without the TestRunner');

var deviceLightEvent;
function checkLight(event) {
    deviceLightEvent = event;
    shouldBe('deviceLightEvent.value', 'mockLight');
}

function listener(event) {
    checkLight(event);
    finishJSTest();
}

window.addEventListener('devicelight', listener);
window.jsTestIsAsync = true;
</script>
</body>
</html>
